bug: pos_processamento() nao esta iterando pelas turmas multiplas vezes

+++ test/main.cpp
@@ -7,7 +7,13 @@ int main(int argc, char **argv)
     auto rd = std::random_device{};
     auto rng = std::default_random_engine{rd()};
 
-    Heuristica h = Heuristica(rng, argv[1], std::stoi(argv[2]), std::stoi(argv[3]), std::stof(argv[4]), std::stof(argv[5]), std::stof(argv[6]));
+    // argv[1] = instancia string
+    // argv[2] = tam pop int
+    // argv[3] = qtd turmas na heuristica int
+    // argv[4] = qtd repeticoes por busca local int
+    // argv[5] = peso janela float
+    // argv[6] =  peso sexto float
+    Heuristica h = Heuristica(rng, argv[1], std::stoi(argv[2]), std::stoi(argv[3]), std::stoi(argv[4]), std::stof(argv[5]), std::stof(argv[6]));
     auto tempo_inicial = std::chrono::steady_clock::now();
     h.inicializar(&tempo_inicial);
